requirements/deploy_requirements.txt
flake8==3.8.3
importlib-metadata==1.6.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.13.0
PyYAML==5.3.1
zipp==3.1.0
